+++ src/Dangl.BCF/BCFv21/BCFv21Container.cs
@@ -251,13 +251,22 @@ namespace Dangl.BCF.BCFv21
                 }
             }
         }
-
         /// <summary>
         ///     Reads a BCFv2 zip archive
         /// </summary>
         /// <param name="zipFileStream">The zip archive of the physical file</param>
         /// <returns></returns>
         public static BCFv21Container ReadStream(Stream zipFileStream)
+        {
+            return ReadStream(zipFileStream, null);
+        }
+
+        /// <summary>
+        ///     Reads a BCFv2 zip archive
+        /// </summary>
+        /// <param name="zipFileStream">The zip archive of the physical file</param>
+        /// <returns></returns>
+        public static BCFv21Container ReadStream(Stream zipFileStream, Action<BCFTopic, System.Xml.Linq.XDocument> topicReadCallback)
         {
             var container = new BCFv21Container();
             var bcfZipArchive = new ZipArchive(zipFileStream, ZipArchiveMode.Read);
@@ -300,7 +309,7 @@ namespace Dangl.BCF.BCFv21
                 {
                     if (!topicIds.Contains(Regex.Match(topicId, @"^\b[A-Fa-f0-9]{8}(?:-[A-Fa-f0-9]{4}){3}-[A-Fa-f0-9]{12}\b").Value))
                     {
-                        container.Topics.Add(ReadSingleTopic(bcfZipArchive, Regex.Match(topicId, @"^\b[A-Fa-f0-9]{8}(?:-[A-Fa-f0-9]{4}){3}-[A-Fa-f0-9]{12}\b").Value, container));
+                        container.Topics.Add(ReadSingleTopic(bcfZipArchive, Regex.Match(topicId, @"^\b[A-Fa-f0-9]{8}(?:-[A-Fa-f0-9]{4}){3}-[A-Fa-f0-9]{12}\b").Value, container, topicReadCallback));
                     }
                     topicIds.Add(Regex.Match(topicId, @"^\b[A-Fa-f0-9]{8}(?:-[A-Fa-f0-9]{4}){3}-[A-Fa-f0-9]{12}\b").Value);
                 }
@@ -404,11 +413,15 @@ namespace Dangl.BCF.BCFv21
             return result.TrimEnd('/');
         }
 
-        private static BCFTopic ReadSingleTopic(ZipArchive archive, string topicId, BCFv21Container container)
+        private static BCFTopic ReadSingleTopic(ZipArchive archive,
+            string topicId,
+            BCFv21Container container,
+            Action<BCFTopic, System.Xml.Linq.XDocument> topicReadCallback)
         {
             var topic = new BCFTopic();
             // Get the markup
-            topic.Markup = Markup.Deserialize(archive.Entries.First(e => e.FullName == topicId + "/" + "markup.bcf").Open());
+            var topicEntry = archive.Entries.First(e => e.FullName == topicId + "/" + "markup.bcf");
+            topic.Markup = Markup.Deserialize(topicEntry.Open());
             // Check if any comments have a Viewpoint object without any value, then set it to null
             foreach (var comment in topic.Markup.Comment.Where(c => c.ShouldSerializeViewpoint() && string.IsNullOrWhiteSpace(c.Viewpoint.Guid)))
             {
@@ -519,6 +532,14 @@ namespace Dangl.BCF.BCFv21
                     }
                 }
             }
+
+            var callback = topicReadCallback;
+            if (callback != null)
+            {
+                var topicDocument = System.Xml.Linq.XDocument.Load(topicEntry.Open());
+                callback(topic, topicDocument);
+            }
+
             return topic;
         }
     }