+++ CHANGELOG.md
@@ -6,3 +6,4 @@ All notable changes to **Dangl.BCF** are documented here.
 
 - Initial release of the **Dangl.BCF** fork
 - Support for BCF v3.0
+- Added a feature for a callback method so users can directly evaluate a BCF XMLs topic XML during import